split keyboard views into separate files
