+++ nuvox_app/keyboard/apps.py
@@ -1,5 +1,9 @@
 from django.apps import AppConfig
 
+from nuvox_algorithm import NuvoxAlgorithm
+
 
 class KeyboardConfig(AppConfig):
     name = 'keyboard'
+
+    # nuvox_algorithm = NuvoxAlgorithm()
