+++ nuvox_app/keyboard/views/__init__.py
@@ -1,0 +1,3 @@
+from .game import game, random_word, DataCollectionSwipeViewSet
+from .keyboard import keyboard
+from .home import home
