+++ nuvox_app/keyboard/views/game.py
@@ -12,17 +12,6 @@ from keyboard.models import DataCollectionSwipe, DeviceType
 from keyboard.serializers import DataCollectionSwipeSerializer
 from keyboard.validators import trace_matches_target_text
 
-
-def home(request):
-    return render(request=request, template_name='home.html')
-
-
-@login_required()
-def keyboard(request):
-    context = {'is_mobile': request.user_agent.is_mobile}
-    return render(request=request, template_name='keyboard/keyboard.html', context=context)
-
-
 @login_required()
 def game(request):
     context = {'is_mobile': request.user_agent.is_mobile}
