+++ nuvox_app/keyboard/views/home.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def home(request):
+    return render(request=request, template_name='home.html')
